Add files via upload
Decision Tree is one of the mostly used algorithm in Machine Learning. Decision Tree is a supervised machine Learning technique. Decision Tree can be used for solving regression as well as classification type of problems. It is used to solve complex decision-making problem.

As the name suggests,the decision tree is structured like an tree, where each internal node represents a attribute (decision), each branch represents an outcome of the attribute, each leaf node represents a class label (in classification) or a numerical value (in regression).